+++ tests/Translator/NodeTranslator/EntityReferenceTranslatorTest.php
@@ -5,7 +5,7 @@ namespace Tests\Innmind\Xml\Translator\NodeTranslator;
 
 use Innmind\Xml\{
     Translator,
-    Node,
+    Element,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -13,15 +13,21 @@ class EntityReferenceTranslatorTest extends TestCase
 {
     public function testTranslate()
     {
+        $document = \Dom\XMLDocument::createFromString($xml = <<<XML
+            <?xml version="1.0" encoding="UTF-8"?>
+            <foo>&gt;</foo>
+            XML
+        );
+
         $translate = Translator::of();
         $node = $translate(
-            new \DOMEntityReference('gt'),
+            $document->childNodes->item(0),
         )->match(
             static fn($node) => $node,
             static fn() => null,
         );
 
-        $this->assertInstanceOf(Node::class, $node);
-        $this->assertSame('gt', $node->content());
+        $this->assertInstanceOf(Element::class, $node);
+        $this->assertSame("<foo>&gt;</foo>\n", $node->asContent()->toString());
     }
 }