+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Requires PHP `8.4`
+- Requires `innmind/immutable:~6.0`
+- Requires `innmind/filesystem:~9.0`
+
 ## 8.0.0 - 2025-06-01
 
 ### Added